desktop_app/requirements.txt
PyQt5==5.15.9
psutil==5.9.5
pynput==1.7.6
pygetwindow==0.0.9
requests==2.31.0
pyjwt==2.8.0
pyinstaller==6.13.0 